+++ packages/@react-elf/ui/src/index.ts
@@ -14,3 +14,4 @@ export * from "@react-elf/toolbar";
 export * from "@react-elf/tools";
 export * from "@react-elf/ghost";
 export * from "@react-elf/badge";
+export * from "@react-elf/panel";