+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @react-elf/docs
 
+## 0.0.76
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @react-elf/ui@0.0.76
+
+## 0.0.75
+
+### Patch Changes
+
+- add Toast, Contextual Help mdx
+
 ## 0.0.74
 
 ### Patch Changes