+++ packages/@react-elf-types/dialog/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @react-elf-types/notification
 
+## 0.0.76
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @react-elf-types/button@0.0.76
+  - @react-elf-types/shared@0.0.76
+
 ## 0.0.73
 
 ### Patch Changes