+++ packages/@react-elf-types/layout/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @react-elf-types/layout
 
+## 0.0.76
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @react-elf-types/shared@0.0.76
+
 ## 0.0.69
 
 ### Patch Changes