+++ packages/@react-elf/badge/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @react-elf/badge
 
+## 0.0.76
+
+### Patch Changes
+
+- add badge, panel component
+
+- Updated dependencies []:
+  - @react-elf/shared@0.0.76
+  - @react-elf-types/badge@0.0.76
+
 ## 0.0.73
 
 ### Patch Changes