+++ packages/@react-elf/toolbar/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @react-elf/toolbar
 
+## 0.0.76
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @react-elf-types/shared@0.0.76
+  - @react-elf/shared@0.0.76
+  - @react-elf-types/toolbar@0.0.76
+  - @react-elf/tools@0.0.76
+
 ## 0.0.71
 
 ### Patch Changes